stripe_test/requirements.txt
asgiref==3.5.2
certifi==2022.6.15.2
charset-normalizer==2.1.1
Django==4.1.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.4
psycopg2-binary==2.9.3
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
stripe==4.1.0
urllib3==1.26.12
whitenoise==5.3.0
